walletdb: Change BerkeleyDatabase::Flush to Close
Change BerkeleyEnvironment::Flush to not shutdown the the environment.
Move the log file cleanup stuff to BerkeleyEnvironment::Close. Also
rename BerkeleyDatabase::Flush to Close as that is really what it is
doing when it calls BerkeleyEnvironment::Flush. The shutdown paramter is
removed and its stuff is moved into BerkeleyDatabase's destructor.