Account for fees when deciding whether to make change
The coin selection algorithm will tell us the fees it expects the
transaction to have. We use these fees to determine whether the
algorithm decided there should be change. This allows us to respect the
no change results.

However we still handle the fee by reducing the change later. This is so
that when we do have a change output, we don't overpay on fees as by
that point, we know the final transaction size and can calculate the
final fees correctly.

Additionally, when doing that reduction, we do the difference between
nFeeRet and fee_needed so that future changes which may account for
nFeeRet earlier will not break the change reduction.