rpc: Be able to provide passphrase to migratewallet
In order to migrate encrypted wallets, the passphrase must be provided,
even if the wallet was loaded and already unlocked. This because the
wallet will be unloaded internally prior to migration.

This is also required for migrating wallets that are not loaded but
instead have their name specified in the RPC.