Add sweep RPC for wallet
Currently, the wallet uses a fSubtractFeeAmount flag on the recipients
objects for all forms of sending calls. The flag was presumably
introduced to permit sweeping without manually calculating the fees of
transactions.  However, the flag leads to unintuitive behavior and makes
it more complicated to test exhaustively many wallet RPCs. I'm proposing
to introduce a dedicated sweep RPC which provides an avenue to easily
empty wallets as a stepping stone towards removing the "subtract from
recipient" option on other wallet RPC.

The proposed sweep call builds a transaction from the wallet's entire
UTXO pool and assigns the funds to one or more recipients. Recipients
can either be specified to receive a given amount or receive an equal
share of the remaining unassigned funds. At least one recipient must be
provided without assigned amount to collect the remainder. The call has
a sendmax option that changes the default behavior of spending all UTXOs
("no UTXO left behind"), to maximizing the output amount of the
transaction by skipping uneconomic UTXOs.