tests: Updates for deprecating SFFO
Several tests were using SFFO in order to sweep the wallet. This usage
has been replaced with the sweep function.

The tests that are testing the SFFO behavior specifically will now have
the necessary nodes restarted with -deprecatedrpc=sffo.

A few tests were using SFFO but did not need to; these have been changed
to not use SFFO.